+++ packages/ng/libraries/core/src/lib/select/input/select-input.component.ts
@@ -47,8 +47,9 @@ import { ALuSelectInput } from './select-input.model';
 export class LuSelectInputComponent<T = any, P extends ILuPickerPanel<T> = ILuPickerPanel<T>>
 extends ALuSelectInput<T, P>
 implements ControlValueAccessor, ILuInputWithPicker<T>, AfterViewInit {
-	protected displayer: ILuInputDisplayer<T>;
-	@ViewChild('display', { read: ViewContainerRef }) protected _displayContainer: ViewContainerRef;
+	@ViewChild('display', { read: ViewContainerRef }) protected set _vcDisplayContainer(vcr: ViewContainerRef) {
+		this.displayContainer = vcr;
+	}
 
 	@Input('placeholder') set inputPlaceholder(p: string) { this._placeholder = p; }
 	@Input('multiple') set inputMultiple(m: boolean | string) {
@@ -85,10 +86,10 @@ implements ControlValueAccessor, ILuInputWithPicker<T>, AfterViewInit {
 	@ContentChild(ALuPickerPanel) set _contentChildPicker(picker: P) {
 		this._picker = picker;
 	}
-	@ContentChild(ALuClearer) set _ContentChildClearer(clearer: ILuClearer) {
+	@ContentChild(ALuClearer) set _contentChildClearer(clearer: ILuClearer) {
 		this._clearer = clearer;
 	}
-	@ContentChild(ALuInputDisplayer) set _displayer(displayer: ILuInputDisplayer<T>) {
+	@ContentChild(ALuInputDisplayer) set _contentChildDisplayer(displayer: ILuInputDisplayer<T>) {
 		this.displayer = displayer;
 		this.render();
 	}
@@ -112,47 +113,6 @@ implements ControlValueAccessor, ILuInputWithPicker<T>, AfterViewInit {
 	@HostListener('blur')
 	onBlur() {
 		super.onBlur();
-	}
-
-	protected render() {
-		if (!this.displayer) { return; }
-		if (this.useMultipleViews()) {
-			this.renderMultipleViews();
-		} else {
-			this.renderSingleView();
-		}
-	}
-	protected useMultipleViews() {
-		return this._multiple  && !this.displayer.multiple;
-	}
-
-	protected renderSingleView() {
-		this.clearDisplay();
-		if (!!this.value) {
-			const newView = this.getView(this.value);
-			this.displayView(newView);
-		}
-	}
-	protected clearDisplay() {
-		this._displayContainer.clear();
-	}
-	protected getView(value: T | T[]) {
-		if (!!this.displayer) {
-			return this.displayer.getViewRef(value);
-		}
-		return undefined;
-	}
-	protected displayView(view) {
-		if (!!view) {
-			this._displayContainer.insert(view);
-		}
-	}
-
-	protected renderMultipleViews() {
-		this.clearDisplay();
-		const values = <T[]>this.value || [];
-		const views = values.map(value => this.getView(value));
-		views.forEach(view => this.displayView(view));
 	}
 
 	ngAfterViewInit() {
