+++ packages/ng/libraries/core/src/lib/user/select/input/user-select-input.component.ts
@@ -45,9 +45,9 @@ import { ALuUserPagedSearcherService, LuUserPagedSearcherService } from '../sear
 export class LuUserSelectInputComponent<U extends IUser = IUser, P extends ILuOptionPickerPanel<U> = ILuOptionPickerPanel<U>>
 extends ALuSelectInput<U, P>
 implements ControlValueAccessor, ILuInputWithPicker<U>, AfterViewInit {
-	protected displayer: ILuInputDisplayer<U>;
-	@ViewChild('display', { read: ViewContainerRef }) protected _displayContainer: ViewContainerRef;
-
+	@ViewChild('display', { read: ViewContainerRef }) protected set _vcDisplayContainer(vcr: ViewContainerRef) {
+		this.displayContainer = vcr;
+	}
 	searchFormat = LuDisplayFullname.lastfirst;
 	@Input('placeholder') set inputPlaceholder(p: string) { this._placeholder = p; }
 	@Input() set fields(fields: string) { this._service.fields = fields; }
@@ -89,10 +89,10 @@ implements ControlValueAccessor, ILuInputWithPicker<U>, AfterViewInit {
 	@ViewChild(ALuPickerPanel) set _vcPicker(picker: P) {
 		this._picker = picker;
 	}
-	@ViewChild(ALuClearer) set _ContentChildClearer(clearer: ILuClearer) {
+	@ViewChild(ALuClearer) set _contentChildClearer(clearer: ILuClearer) {
 		this._clearer = clearer;
 	}
-	@ViewChild(ALuInputDisplayer) set _displayer(displayer: ILuInputDisplayer<U>) {
+	@ViewChild(ALuInputDisplayer) set _ccDisplayer(displayer: ILuInputDisplayer<U>) {
 		this.displayer = displayer;
 		this.render();
 	}
@@ -118,48 +118,6 @@ implements ControlValueAccessor, ILuInputWithPicker<U>, AfterViewInit {
 	@HostListener('blur')
 	onBlur() {
 		super.onBlur();
-	}
-
-
-	protected render() {
-		if (!this.displayer) { return; }
-		if (this.useMultipleViews()) {
-			this.renderMultipleViews();
-		} else {
-			this.renderSingleView();
-		}
-	}
-	protected useMultipleViews() {
-		return this._multiple  && !this.displayer.multiple;
-	}
-
-	protected renderSingleView() {
-		this.clearDisplay();
-		if (!!this.value) {
-			const newView = this.getView(this.value);
-			this.displayView(newView);
-		}
-	}
-	protected clearDisplay() {
-		this._displayContainer.clear();
-	}
-	protected getView(value: U | U[]) {
-		if (!!this.displayer) {
-			return this.displayer.getViewRef(value);
-		}
-		return undefined;
-	}
-	protected displayView(view) {
-		if (!!view) {
-			this._displayContainer.insert(view);
-		}
-	}
-
-	protected renderMultipleViews() {
-		this.clearDisplay();
-		const values = <U[]>this.value || [];
-		const views = values.map(value => this.getView(value));
-		views.forEach(view => this.displayView(view));
 	}
 
 	ngAfterViewInit() {
